move logic of checkdata from services to postdata
